Revert "enh(UI): Make the reset password form customizable (#6056)"
This reverts commit faa1ae0af7c973f70747f9471a1b0a29ef54f005.
