+++ centreon/www/front_src/src/Login/index.tsx
@@ -22,7 +22,6 @@ import {
   labelLogin,
   labelPoweredByCentreon
 } from './translatedLabels';
-import useGetLoginCustomData from './useGetLoginCustomData';
 import useLogin from './useLogin';
 import useValidationSchema from './validationSchema';
 
@@ -73,11 +72,10 @@ const LoginPage = (): JSX.Element => {
   const { t } = useTranslation();
   const validationSchema = useValidationSchema();
 
-  const { loginPageCustomisation } = useGetLoginCustomData();
-
   const {
     submitLoginForm,
     providersConfiguration,
+    loginPageCustomisation,
     authenticationError,
     hasForcedProvider
   } = useLogin();
