+++ centreon/www/front_src/src/Login/useLogin.ts
@@ -1,8 +1,9 @@
 import { useCallback, useEffect } from 'react';
 
 import { FormikHelpers, FormikValues } from 'formik';
-import { useAtom, useSetAtom } from 'jotai';
+import { useAtom, useAtomValue, useSetAtom } from 'jotai';
 import {
+  path,
   equals,
   filter,
   isEmpty,
@@ -17,18 +18,27 @@ import { useTranslation } from 'react-i18next';
 import { useNavigate, useSearchParams } from 'react-router';
 
 import { ResponseError, useFetchQuery, useSnackbar } from '@centreon/ui';
+import { platformVersionsAtom } from '@centreon/ui-context';
 
 import { platformInstallationStatusAtom } from '../Main/atoms/platformInstallationStatusAtom';
 import useInitializeTranslation from '../Main/useInitializeTranslation';
 import useUser from '../Main/useUser';
 import { passwordResetInformationsAtom } from '../ResetPassword/passwordResetInformationsAtom';
 import { PlatformInstallationStatus } from '../api/models';
+import centreonLogo from '../assets/logo-centreon-colors.svg';
 import routeMap from '../reactRoutes/routeMap';
 
-import { providersConfigurationDecoder } from './api/decoder';
-import { providersConfigurationEndpoint } from './api/endpoint';
+import {
+  loginPageCustomisationDecoder,
+  providersConfigurationDecoder
+} from './api/decoder';
+import {
+  loginPageCustomisationEndpoint,
+  providersConfigurationEndpoint
+} from './api/endpoint';
 import {
   LoginFormValues,
+  LoginPageCustomisation,
   ProviderConfiguration,
   Redirect,
   RedirectAPI
@@ -38,10 +48,12 @@ import {
   labelPasswordHasExpired
 } from './translatedLabels';
 import usePostLogin from './usePostLogin';
+import useWallpaper from './useWallpaper';
 
 interface UseLoginState {
   authenticationError: string | null;
   hasForcedProvider: boolean;
+  loginPageCustomisation: LoginPageCustomisation;
   platformInstallationStatus: PlatformInstallationStatus | null;
   providersConfiguration: Array<ProviderConfiguration> | null;
   submitLoginForm: (
@@ -65,6 +77,14 @@ const getActiveProviders = filter<ProviderConfiguration>(
   propEq(true, 'isActive')
 );
 
+const defaultLoginPageCustomisation: LoginPageCustomisation = {
+  customText: null,
+  iconSource: centreonLogo,
+  imageSource: null,
+  platformName: null,
+  textPosition: null
+};
+
 export const router = {
   useNavigate,
   useSearchParams
@@ -87,6 +107,22 @@ const useLogin = (): UseLoginState => {
     }
   });
   const [platformInstallationStatus] = useAtom(platformInstallationStatusAtom);
+  const platformVersions = useAtomValue(platformVersionsAtom);
+  const { data: loginPageCustomisationData, isFetching } =
+    useFetchQuery<LoginPageCustomisation>({
+      decoder: loginPageCustomisationDecoder,
+      getEndpoint: () => loginPageCustomisationEndpoint,
+      getQueryKey: () => ['loginPageCustomisation'],
+      httpCodesBypassErrorSnackbar: [404, 401],
+      queryOptions: {
+        enabled: !!path(
+          ['modules', 'centreon-it-edition-extensions'],
+          platformVersions
+        ),
+        retry: false,
+        suspense: false
+      }
+    });
 
   const { getInternalTranslation, getExternalTranslation } =
     useInitializeTranslation();
@@ -171,7 +207,27 @@ const useLogin = (): UseLoginState => {
 
   const activeProviders = getActiveProviders(externalProviders || []);
 
+  const wallpaper = useWallpaper();
+
   const authenticationError = searchParams.get('authenticationError');
+
+  const loginPageCustomisation = isFetching
+    ? defaultLoginPageCustomisation
+    : {
+        customText:
+          loginPageCustomisationData?.customText ||
+          defaultLoginPageCustomisation.customText,
+        iconSource:
+          loginPageCustomisationData?.iconSource ||
+          defaultLoginPageCustomisation.iconSource,
+        imageSource: loginPageCustomisationData?.imageSource || wallpaper,
+        platformName:
+          loginPageCustomisationData?.platformName ||
+          defaultLoginPageCustomisation.platformName,
+        textPosition:
+          loginPageCustomisationData?.textPosition ||
+          defaultLoginPageCustomisation.textPosition
+      };
 
   useEffect(() => {
     if (!prop('REDIRECT_URI', cookies)) {
@@ -191,6 +247,7 @@ const useLogin = (): UseLoginState => {
   return {
     authenticationError,
     hasForcedProvider: !!forcedProviders,
+    loginPageCustomisation,
     platformInstallationStatus,
     providersConfiguration: activeProviders,
     submitLoginForm
